pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "graphsast"
version = "0.1.0"
description = "Security analysis tool combining a structural code graph, vulnerability pattern DB, and an agentic LLM layer"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0,<2",
    "fastmcp>=0.1.0,<2",
    "tree-sitter>=0.23.0,<1",
    "tree-sitter-language-pack>=0.3.0,<1",
    "networkx>=3.2,<4",
    "watchdog>=4.0.0,<6",
    "typer>=0.12,<1",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
    "ruff>=0.3.0,<1",
]

[project.scripts]
graphsast = "graphsast.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["graphsast", "code_review_graph"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["tests/fixtures"]
